Initial commit of RefET-GEE code
This was copied from RefET v0.2.4 and then converted to GEE equations.
Currently all of the calculation functions are passing for ee.Number
input types.  The main daily and hourly functions have not been tested.
Most of the functionality should be moved into a RefET class and methods
for reading common image collections should be created (GRIDMET, NLDAS,
CIMIS, etc).