All functions work for ee.Number inputs
Making the ASCE method be the default for all functions.

It is inconsistent that the daily and hourly are in separate files, but
the tests for both are combined in test_refet.

Still need to test the gridmet staticmethod, test that the functions
with image inputs, decide whether it is importantt that the functions
work for ee.Number inputs, and figure out how to have the class convert
the inputs from floats to ee.Number automatically.