Added initials support for RTMA
I added both Hourly and Daily functions.  I'm not 100% sure I have the units on the solar setup but everything runs.  Still need to decide exactly what the default solar should be if it is not set, should it have a default if not set, and how to structure the solar input (image or collection).